chore: 소스코드 위치 변경
- 적절한 위치로 파일 옮김
- gitignore에 yml추가
