--- src/main/resources/application.yml
@@ -1,35 +1,0 @@
-spring:
-  mvc:
-    favicon:
-      enabled:false
-  application:
-    name: chzzkzzal
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:testdb;MODE=MySQL;DB_CLOSE_DELAY=-1
-    username: sa
-    password:
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.H2Dialect
-        format_sql: true
-    database-platform: org.hibernate.dialect.H2Dialect
-
-# 하루 = 86400
-session:
-  timeout: 86400
-
-# 쿠키 시간 늘리기
-server:
-  reactive:
-    session:
-      cookie:
-        max-age: 86400
